extract carousel component
