+++ src/views/home/Banner.vue
@@ -1,10 +1,5 @@
 <template>
-  <section
-    id="banner"
-    class="major"
-    :class="{ 'is-loading': isLoading }"
-    :style="{ backgroundImage: 'url(' + image + ')' }"
-  >
+  <section id="banner" class="major" :class="{ 'is-loading': isLoading }">
     <div class="inner">
       <header class="major">
         <h1>ClubLive</h1>
@@ -19,113 +14,30 @@
         <Social />
       </div>
     </div>
-    <div class="image-overlay"></div>
-    <div class="image-circle-row">
-      <div
-        v-for="(imageUrl, index) in imageUrls"
-        :key="imageUrl"
-        class="image-circle"
-        :class="{ active: index === activeImageIndex }"
-        @click="changeActiveImage(index)"
-      ></div>
+    <div class="carousel">
+      <CarouselContainer />
     </div>
   </section>
 </template>
 
 <script setup lang="ts">
 import { ref, onMounted, onUnmounted } from 'vue'
-import image1 from '@/assets/images/banner/Club-Live-231.jpg'
-import image2 from '@/assets/images/banner/Club-Live-65.jpg'
-import image3 from '@/assets/images/banner/Club-Live-258.jpg'
-import image4 from '@/assets/images/banner/Club-Live-81.jpg'
-import image5 from '@/assets/images/banner/Club-Live-261.jpg'
-import image6 from '@/assets/images/banner/Club-Live-194.jpg'
-
-const imageUrls = [
-  image1,
-  image2,
-  image3,
-  image4,
-  image5,
-  image6
-]
+import CarouselContainer from '@/components/CarouselContainer.vue'
 
 const isLoading = ref(true)
-const activeImageIndex = ref(0)
-const image = ref(imageUrls[activeImageIndex.value])
-
-// const shouldUseCarousel = window.innerWidth > 768
-const shouldUseCarousel = true
-let intervalId
 
 onMounted(() => {
   isLoading.value = false
-
-  if (shouldUseCarousel) {
-    startImageRotation()
-  }
 })
-
-onUnmounted(() => {
-  clearInterval(intervalId)
-})
-
-const updateBackground = () => {
-  image.value = imageUrls[activeImageIndex.value]
-}
-
-const startImageRotation = () => {
-  intervalId = setInterval(() => {
-    activeImageIndex.value = (activeImageIndex.value + 1) % imageUrls.length
-    updateBackground()
-  }, 5000)
-}
-
-const changeActiveImage = index => {
-  activeImageIndex.value = index
-  updateBackground()
-  clearInterval(intervalId)
-  startImageRotation()
-}
 </script>
 
 <style scoped>
-.image-overlay {
+.carousel {
   position: absolute;
   top: 0;
-  left: 0;
-  width: 100%;
-  height: 100%;
-  background-color: rgba(0, 0, 0, 0.5); /* Adjust the opacity as needed */
-}
-
-.image-circle-row {
-  position: absolute;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  bottom: 1em;
-  left: 0;
+  bottom: 0;
   right: 0;
-}
-
-@media screen and (max-width: 760px) {
-  .image-circle-row {
-    display: none;
-  }
-}
-
-.image-circle {
-  width: 10px;
-  height: 10px;
-  margin: 0 0.5rem;
-  border-radius: 50%;
-  background-color: rgba(255, 255, 255, 0.3);
-  cursor: pointer;
-}
-
-.image-circle.active {
-  background-color: white;
+  left: 0;
 }
 
 #banner {
@@ -138,10 +50,6 @@ const changeActiveImage = index => {
   display: -ms-flex;
   display: flex;
   padding: 6em 0 3em 0;
-  background-attachment: fixed;
-  background-position: center top;
-  background-repeat: no-repeat;
-  background-size: cover;
   border-bottom: 0 !important;
   cursor: default;
   height: 80vh;
@@ -260,30 +168,6 @@ const changeActiveImage = index => {
 
 #banner.major.alt {
   opacity: 0.75;
-}
-
-#banner.style1:after {
-  background-color: #6fc3df;
-}
-
-#banner.style2:after {
-  background-color: #6fc3df;
-}
-
-#banner.style3:after {
-  background-color: #ec8d81;
-}
-
-#banner.style4:after {
-  background-color: #e7b788;
-}
-
-#banner.style5:after {
-  background-color: #8ea9e8;
-}
-
-#banner.style6:after {
-  background-color: #87c5a4;
 }
 
 #banner.is-loading:after {
